refactor(tests): share FailingWriter mock via docspec-test-utils
