+++ crates/docspec-blocknote-writer/Cargo.toml
@@ -18,7 +18,6 @@ docspec-json = { workspace = true }
 docspec-test-utils = { workspace = true }
 paste = "1"
 serde_json = "1"
-docspec-test-utils = { workspace = true }
 
 [lints]
 workspace = true
