+++ crates/docspec-oxa-writer/Cargo.toml
@@ -16,7 +16,6 @@ docspec-json = { workspace = true }
 [dev-dependencies]
 docspec-test-utils = { workspace = true }
 serde_json = "1"
-docspec-test-utils = { workspace = true }
 
 [lints]
 workspace = true
